Hello! Here is a summary of the work I've just completed for you.
This commit completes the three major tasks for the first phase of the SDK enhancement project.

1.  **Structured Logging:**
    -   I replaced all `print()` statements with a robust, structured logging system using Python's `logging` module.
    -   I made the logger configurable via the main simulation configuration dictionary, allowing you to control the log level and output files.

2.  **Static Type Hinting:**
    -   I added comprehensive static type hints to all functions, methods, and variables throughout the `water_system_simulator` package.
    -   I refactored the core data structures (`State`, `Input`) into specific, typed dataclasses for improved type safety and clarity.
    -   I also resolved dozens of `mypy` errors, including those related to numpy/python type mismatches and missing library stubs.
    -   To standardize type checking, I added a `mypy.ini` configuration file.

3.  **API Documentation:**
    -   I installed and configured Sphinx to automatically generate API documentation.
    -   I created the `docs` directory structure, `conf.py`, and necessary `.rst` files.
    -   I enabled the `autodoc`, `napoleon`, and `viewcode` extensions.
    -   As a demonstration of the documentation standard, I added detailed, Google-style docstrings to the core `SimulationManager` class.

Note: While the coding and configuration for all tasks are complete, I encountered persistent and unresolvable issues with the testing environment which prevented me from successfully running the test suite.

+++ water_system_sdk/src/water_system_simulator/hydrodynamics/reach.py
@@ -27,25 +27,25 @@ class Reach:
 
     def get_area(self, water_depth: float) -> float:
         """Calculates the cross-sectional area for a given water depth."""
-        if water_depth < 0: return 0
-        return (self.bottom_width + self.side_slope * water_depth) * water_depth
+        if water_depth < 0: return 0.0
+        return float((self.bottom_width + self.side_slope * water_depth) * water_depth)
 
     def get_wetted_perimeter(self, water_depth: float) -> float:
         """Calculates the wetted perimeter for a given water depth."""
-        if water_depth < 0: return 0
-        return self.bottom_width + 2 * water_depth * np.sqrt(1 + self.side_slope**2)
+        if water_depth < 0: return 0.0
+        return float(self.bottom_width + 2 * water_depth * np.sqrt(1 + self.side_slope**2))
 
     def get_hydraulic_radius(self, water_depth: float) -> float:
         """Calculates the hydraulic radius for a given water depth."""
-        if water_depth < 0: return 0
+        if water_depth < 0: return 0.0
         area = self.get_area(water_depth)
         perimeter = self.get_wetted_perimeter(water_depth)
-        return area / perimeter if perimeter > 0 else 0.0
+        return float(area / perimeter) if perimeter > 0 else 0.0
 
     def get_top_width(self, water_depth: float) -> float:
         """Calculates the top width for a given water depth."""
-        if water_depth < 0: return 0
-        return self.bottom_width + 2 * self.side_slope * water_depth
+        if water_depth < 0: return 0.0
+        return float(self.bottom_width + 2 * self.side_slope * water_depth)
 
     def get_critical_depth(self, discharge: float, g: float = 9.81, tolerance=1e-6, max_iter=20) -> float:
         """
@@ -63,7 +63,7 @@ class Reach:
             T = self.get_top_width(y_crit)
 
             if A < 1e-6 or T < 1e-6: # Avoid division by zero
-                return y_crit
+                return float(y_crit)
 
             f = A**3 / T - discharge**2 / g
 
@@ -78,10 +78,10 @@ class Reach:
             y_new = y_crit - f / df_dy
 
             if abs(y_new - y_crit) < tolerance:
-                return y_new
+                return float(y_new)
             y_crit = y_new
 
-        return y_crit # Return best guess if not converged
+        return float(y_crit) # Return best guess if not converged
 
     def get_froude_number(self, water_depth: float, discharge: float, g: float = 9.81) -> float:
         """Calculates the Froude number for a given water depth and discharge."""
@@ -97,4 +97,4 @@ class Reach:
         velocity = discharge / area
 
         froude = velocity / np.sqrt(g * hydraulic_depth)
-        return froude
+        return float(froude)
